refactor(web): AssigneeDropdown を features/common へ移動し feature 間の横断依存を解消する
